[ FIX ] Long to int

+++ android/player/src/main/java/com/ghondar/gplayer/GPlayer.java
@@ -87,7 +87,7 @@ public class GPlayer extends ReactContextBaseJavaModule {
     }
 
     @ReactMethod
-    public void setDefaultRetryTime(Long defaultRetryTime) {
+    public void setDefaultRetryTime(int defaultRetryTime) {
         config.setDefaultRetryTime(defaultRetryTime);
     }
